@@ -1,5 +1,5 @@
-module github.com/The-Cloud-Source/pureapi/pure1
+module github.com/the-cloud-source/pureapi/pure1
 
-go 1.12
+go 1.18
 
 require github.com/dgrijalva/jwt-go v3.2.0+incompatible
